Optimize code for readability not speed.
We will need to program 4 different ESCs, meaning we will need to switch easily between pins. This is hard to do when writing directly to port regs
Clock needs to be pulsed using bit manipulation for some reason. I can not get the timing to be correct using arduino digitalwrite functions"