Concat hex file and send bigger buffers where possible.
When the lines of the hex file should be written directly after eachother in the ESC mem these can be concatenated for speed optimization.
Somtimes the lines should not be next to eachother in mem. In that case we should write the part that can be concatenated before writing the
data that should go in other parts of mem